# Conflicts:
#	app/build.gradle.kts

+++ app/src/main/java/com/example/mimascota/Model/Alimento.kt
@@ -4,7 +4,7 @@ data class Alimento(
     val idPro: String,
     val namePro: String,
     val descriptionPro: String,
-    val pricePro: Double,
+    val pricePro: Int,
     val marca: String,
     val tipo: String,
     val peso: Double
