+++ app/src/main/java/com/example/mimascota/Model/Juguete.kt
@@ -4,7 +4,7 @@ data class Juguete(
     val idPro: String,
     val namePro: String,
     val descriptionPro: String,
-    val pricePro: Double,
+    val pricePro: Int,
     val material: String,
     val tamano: String
 ) : Product(idPro, namePro, descriptionPro, pricePro) {
